requirements.txt
langchain==0.0.281
chromadb==0.4.7
llama-cpp-python==0.1.81
urllib3==2.0.4
python-dotenv==1.0.0
unstructured==0.10.8
extract-msg==0.45.0
tqdm==4.66.1
sentence_transformers==2.2.2
torch==2.0.1